[chore] split webpack config for modularity
