+++ config/webpack.copyfiles.config.js
@@ -1,0 +1,11 @@
+'use strict'
+
+const CopyPlugin = require('copy-webpack-plugin')
+
+module.exports = {
+  plugins: [
+    new CopyPlugin([
+      { from: 'vendor/assets', ignore: ['.gitkeep'] }
+    ])
+  ]
+}
